feat: add currency visual transformation

+++ app/src/main/java/com/alexcao/dexpense/data/models/ExpenseInfo.kt
@@ -5,6 +5,7 @@ import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 
 @Entity(
@@ -31,7 +32,7 @@ data class ExpenseInfo(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     val label: String = "",
-    val amount: BigDecimal = BigDecimal.ZERO,
+    val amount: BigInteger = BigInteger.ZERO,
     val date: LocalDate = LocalDate.now(),
     val categoryId: Int = 0,
     val expenseSourceId: Int = 0,
